refactor: cra to vite migration
