+++ my-app/src/Components/Stars.js
