+++ my-app/src/assets/loader.gif
