+++ my-app/src/pages/Detail.js
