+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "renewal-nomflix",
-  "version": "0.1.0",
+  "name": "flix-stack",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^1.2.30",
@@ -45,5 +45,6 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  }
+  },
+  "packageManager": "pnpm@9.7.1"
 }
